REF: moving around functions
